backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
flask-jwt-extended
python-dotenv
openai
PyPDF2
psycopg2-binary
werkzeug
